refactor(bar): use native apis for positioning
This commit replaces almost all uses of the egui API for bar window
positioning with calls to SetWindowPos via komorebi_client's Window
struct.

This seems to play much more smoothly with multi-monitor setups where
each monitor has a different scaling factor, opening the door for
multiple instances of komorebi-bar.exe to run against multiple monitors.

As a result of this change, the "viewport" configuration option has been
renamed to "position" and doc strings have been changed to remove the
reference to the egui crate docs. Similarly, "viewport.position" and
"viewport.inner_size" have been renamed to "position.start" and
"position.end" respectively. Backwards-compatibility aliases have been
included for all renames.

For now, there is one breaking configuration change which I am going to
try to find a way to either walk back or provide an automated
remediation for on first-launch before v0.1.30 is released: both "start"
and "end" are now required sub-options for "position", as we are calling
a Win32 API which takes a Rect and needs all four points.